+++ SampleCode/WorkWallet.BI.ClientDatabaseDeploy/Scripts/StoredProcedures/mart.ETL_DeleteReportedIssueFacts.sql
@@ -41,6 +41,13 @@ BEGIN
 
     PRINT 'DELETE mart.ReportedIssuePersonFact, number of rows = ' + CAST(@@ROWCOUNT AS varchar);
 
+    DELETE mart.ReportedIssuePersonFact2
+    FROM mart.ReportedIssuePersonFact2 AS d
+    INNER JOIN mart.ReportedIssue AS ri ON d.ReportedIssue_key = ri.ReportedIssue_key
+    INNER JOIN @reportedIssueTable AS a ON ri.ReportedIssueId = a.ReportedIssueId;
+
+    PRINT 'DELETE mart.ReportedIssuePersonFact2, number of rows = ' + CAST(@@ROWCOUNT AS varchar);
+
     DELETE mart.ReportedIssueInvestigationTeamFact
     FROM mart.ReportedIssueInvestigationTeamFact AS d
     INNER JOIN mart.ReportedIssue AS ri ON d.ReportedIssue_key = ri.ReportedIssue_key